API BREAK: libxdot: use size_t instead of int for stats and polygon line points
This allows exceeding counts of 2³¹ - 1, which is not particularly likely. But
its more important effect is to make arithmetic operations less error prone and
make it impossible by-construction to arrive at a negative count in these fields.

There is some interaction between the stats fields and polygon line points
fields, so doing these two separately would have meant an intermediate state
with lots of casting. It seemed simpler and less error prone to do this all at
once.

There is still casting back to int for the `gvrender_*` functions. Modifying
these would have been too invasive. Surprisingly the call to `gvrender_polyline`
affected in this commit is the only instance of a call to it with a non-literal
point count. But modifying it to take a size_t would still have required
too-invasive changes due to other functions it calls.

For lines that would have been changed anyway, this commit also fixes white
space and tightens variable scoping where possible.

+++ lib/xdot/xdot.h
@@ -78,7 +78,7 @@ typedef struct {
 } xdot_rect;
 
 typedef struct {
-    int cnt;
+    size_t cnt;
     xdot_point* pts;
 } xdot_polyline;
 
@@ -153,20 +153,20 @@ typedef struct {
 
 typedef struct {
     size_t cnt;  /* no. of xdot ops */
-    int n_ellipse;
-    int n_polygon;
-    int n_polygon_pts;
-    int n_polyline;
-    int n_polyline_pts;
-    int n_bezier;
-    int n_bezier_pts;
-    int n_text;
-    int n_font;
-    int n_style;
-    int n_color;
-    int n_image;
-    int n_gradcolor;
-    int n_fontchar;
+    size_t n_ellipse;
+    size_t n_polygon;
+    size_t n_polygon_pts;
+    size_t n_polyline;
+    size_t n_polyline_pts;
+    size_t n_bezier;
+    size_t n_bezier_pts;
+    size_t n_text;
+    size_t n_font;
+    size_t n_style;
+    size_t n_color;
+    size_t n_image;
+    size_t n_gradcolor;
+    size_t n_fontchar;
 } xdot_stats;
 
 /* ops are indexed by xop_kind */
